Make deploy.sh smarter, add machine-specific config to config.fish, add vim-plist

+++ .config/fish/config.fish
@@ -4,13 +4,25 @@ export EDITOR='vim'
 
 alias svim='sudo vim'
 alias chkjrn='sudo journalctl -xe'
-alias xmap='sh ~/.config/xkbcomp/vim-keys-xkb.sh'
-alias startfusuma='killall fusuma && fusuma -d'
-alias load-kvm="sudo /home/mike/Documents/scripts/intel_gvt-g_setup/start-kvm-gvt-g.sh"
 alias ls="exa"
 alias cat="bat"
+
 
 # Launch tmux
 if test ! -n "$TMUX"
   tmux
+end
+
+# Personal Machine Setup
+if test (uname -n) = kratos
+  alias xmap='sh ~/.config/xkbcomp/vim-keys-xkb.sh'
+  alias startfusuma='killall fusuma && fusuma -d'
+  alias load-kvm="sudo /home/mike/Documents/scripts/intel_gvt-g_setup/start-kvm-gvt-g.sh"
+end
+
+# Work Machine Setup
+if test (uname -n) = TD-C02FK3H8MD6T 
+  export KUBECONFIG="$HOME/.kube/config:$HOME/.kube/config-dev"
+  alias kdev='kubectl config use-context appsec-dev'
+  alias kprod='kubectl config use-context app-sec'
 end
